+++ include/info.h
@@ -3343,12 +3343,12 @@ NUMSTATES
 
 typedef struct
 {
-	 spritenum_t	sprite;
-	 long			frame;
-	 long			tics;
-	 void			(*action) ();
-	 statenum_t		nextstate;
-	 long			misc1, misc2;
+	spritenum_t	sprite;
+	int32_t			frame;
+	int32_t			tics;
+	void			(*action) ();
+	statenum_t		nextstate;
+	int32_t			misc1, misc2;
 } state_t;
 
 extern state_t	states[NUMSTATES];
